ครบ หมด เหลือแค่ ดาต้ากริดวิว สองตาราง

+++ GCSViews/Form_farm_Schedule.cs
@@ -58,7 +58,7 @@ namespace MissionPlanner.GCSViews
             //string a = "yyyy-MM-dd";
             string dateNow = String.Format("{0:yyyy-MM-dd}", DateTime.Now);
 
-            String query = "SELECT * FROM schedule_action WHERE farm_id = '" + id_farm + "' AND action_date >= dateNow";
+            String query = "SELECT * FROM schedule_action WHERE farm_id = '" + id_farm + "' AND action_date >= '" + dateNow + "' ";
             SqlDataAdapter SDA = new SqlDataAdapter(query, con);
             DataTable dt = new DataTable();
             SDA.Fill(dt);
@@ -69,11 +69,14 @@ namespace MissionPlanner.GCSViews
         private void button_delete_Click(object sender, EventArgs e)
         {
             // delete Schedule
-            con.Open();
-            String query = "DELETE FROM schedule_action where action_id = '" + textBox_actID.Text + "' ";
-            SqlDataAdapter SDA = new SqlDataAdapter(query, con);
-            SDA.SelectCommand.ExecuteNonQuery();
-            con.Close();
+            if (MessageBox.Show("Are you wnat to delete this record?", "Message", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            {
+                con.Open();
+                String query = "DELETE FROM schedule_action where action_id = '" + textBox_actID.Text + "' ";
+                SqlDataAdapter SDA = new SqlDataAdapter(query, con);
+                SDA.SelectCommand.ExecuteNonQuery();
+                con.Close();
+            }
             MessageBox.Show("DELETE Record From DB Success!!");
         }
 
@@ -93,6 +96,26 @@ namespace MissionPlanner.GCSViews
         {
             // TODO: This line of code loads data into the 'schedule_act.schedule_action' table. You can move, or remove it, as needed.
             this.schedule_actionTableAdapter.Fill(this.schedule_act.schedule_action);
+
+        }
+
+        private void textBox_actName_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void textBox_actID_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label2_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label1_Click(object sender, EventArgs e)
+        {
 
         }
     }
